+++ src/app/mypage/page.tsx
@@ -1,15 +1,13 @@
 'use client';
 
 import { useState } from 'react';
-import Image from 'next/image';
 
 import Button from '@/components/@shared/Button';
-import Modal from '@/components/@shared/Modal';
-import Input from '@/components/@shared/Input';
 
-import MyGathering from './myGathering';
-import MyLike from './myLike';
-import MyReview from './myReview';
+import MyGathering from '../../components/mypage/myGathering';
+import MyLike from '../../components/mypage/myLike';
+import MyReview from '../../components/mypage/myReview';
+import MyProfileEditModal from '../../components/mypage/myProfileEditModal';
 
 export default function MyPage() {
   const user = {
@@ -22,13 +20,8 @@ export default function MyPage() {
   };
 
   const [isModal, setIsModal] = useState(false);
-  const [img, setImg] = useState<string | null>(null);
 
   const openModalhandler = () => setIsModal(true);
-  const closeModalhandler = () => {
-    setIsModal(false);
-    setImg(null);
-  };
 
   const navLinks = [
     { label: '나의 모임', component: <MyGathering /> },
@@ -45,33 +38,6 @@ export default function MyPage() {
   const renderActiveComponent = () => {
     const activeLink = navLinks.find((link) => link.label === activeTab);
     return activeLink?.component;
-  };
-
-  const handleFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-
-    const isEnglishName = /^[a-zA-Z0-9._-]+$/.test(file.name);
-    if (!isEnglishName) {
-      alert('파일 이름은 영어로만 이루어져야 합니다.');
-      return;
-    }
-
-    const maxSizeInBytes = 5 * 1024 * 1024;
-    if (file.size > maxSizeInBytes) {
-      alert('파일 크기는 5MB 이하여야만 합니다.');
-      return;
-    }
-
-    const previewUrl = URL.createObjectURL(file);
-    setImg(previewUrl);
-  };
-
-  const triggerFileInput = () => {
-    const fileInput = document.getElementById(
-      'profile_image-input'
-    ) as HTMLInputElement;
-    if (fileInput) fileInput.click();
   };
 
   return (
@@ -92,60 +58,12 @@ export default function MyPage() {
         >
           프로필 편집
         </Button>
-        <Modal
-          isOpen={isModal}
-          onClose={closeModalhandler}
-          customDimStyle="w-[400px]"
-        >
-          <div className="flex flex-col gap-5">
-            <p className="text-base font-bold">프로필 수정하기</p>
-            <div className="h-[65px] w-[65px] rounded-full">
-              <div
-                className={`relative flex items-center justify-center overflow-hidden ${
-                  img ? 'w-65 h-65 rounded-full' : ''
-                }`}
-              >
-                <Image
-                  src={img || '/profile_image_default.png'}
-                  width={65}
-                  height={65}
-                  alt="프로필 이미지 미리보기"
-                  className={img ? 'h-[65px] rounded-full' : ''}
-                />
-              </div>
-              <button type="button" onClick={triggerFileInput}>
-                <Image
-                  src="/edit.png"
-                  width={30}
-                  height={30}
-                  alt="프로필이미지 수정 버튼 이미지"
-                  className="absolute bottom-[180px] left-[65px]"
-                />
-              </button>
-              <input
-                type="file"
-                id="profile_image-input"
-                accept="image/*"
-                className="hidden"
-                onChange={handleFileChange}
-              />
-            </div>
-            <Input
-              placeholder={user.nickname}
-              label="닉네임"
-              fontSize="14"
-              gap="8"
-            />
-            <div className="flex justify-center gap-3">
-              <Button variant="secondary" size="large" font="14">
-                취소하기
-              </Button>
-              <Button variant="primary" size="large" font="14" disabled>
-                수정하기
-              </Button>
-            </div>
-          </div>
-        </Modal>
+        <MyProfileEditModal
+          isModal={isModal}
+          setIsModal={setIsModal}
+          nickname={user.nickname}
+          image={user.image}
+        />
       </div>
       <div className="mx-12">
         <nav className="flex gap-5">
@@ -154,7 +72,7 @@ export default function MyPage() {
               key={link.label}
               type="button"
               className={`text-lg font-bold ${
-                activeTab === link.label ? 'border-b-2 border-black' : ''
+                activeTab === link.label ? 'border-b-2 border-white' : ''
               }`}
               onClick={() => navClick(link.label)}
             >
